filterx/func-timestamp: rename files to accomodate both set/get
