+++ lib/filterx/CMakeLists.txt
@@ -51,7 +51,7 @@ set(FILTERX_HEADERS
     filterx/func-sdata.h
     filterx/func-set-fields.h
     filterx/func-set-pri.h
-    filterx/func-set-timestamp.h
+    filterx/func-timestamp.h
     filterx/func-str-transform.h
     filterx/func-str.h
     filterx/func-unset-empties.h
@@ -124,7 +124,7 @@ set(FILTERX_SOURCES
     filterx/func-sdata.c
     filterx/func-set-fields.c
     filterx/func-set-pri.c
-    filterx/func-set-timestamp.c
+    filterx/func-timestamp.c
     filterx/func-str-transform.c
     filterx/func-str.c
     filterx/func-unset-empties.c
